create TeacherViewingStudentProjectAlert.tsx component
