+++ apps/src/lab2/views/alerts/teacherViewingStudentProject/index.ts
@@ -1,0 +1,1 @@
+export {default} from './TeacherViewingStudentProjectAlert';
