+++ apps/src/lab2/views/alerts/teacherViewingStudentProject/teacherViewingStudentProjectAlert.module.scss
@@ -1,0 +1,10 @@
+@use '@cdo/apps/lab2/views/components/layout/variables.scss' as variables;
+
+// Overrides styles for the Alert component to
+// make it fit better in the Workspace area
+.alertBanner {
+  border-bottom: variables.$default-border;
+  border-radius: 0;
+  padding-block: 0;
+  height: 39px;
+}
